+++ src/main/java/com/esliceu/movies/services/LanguageRoleServices.java
@@ -3,6 +3,7 @@ package com.esliceu.movies.services;
 
 import com.esliceu.movies.models.Language;
 import com.esliceu.movies.models.LanguageRole;
+import com.esliceu.movies.models.Person;
 import com.esliceu.movies.repos.LanguageRoleRepo;
 import com.google.gson.Gson;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,8 @@ import java.util.stream.Collectors;
 public class LanguageRoleServices {
     @Autowired
     LanguageRoleRepo languageRoleRepo;
+    @Autowired
+    PermissionsServices permissionsServices;
 
     public Page<LanguageRole> findAllLanguageRoles(Pageable pageable) {
         return languageRoleRepo.findAll(pageable);
@@ -34,20 +37,6 @@ public class LanguageRoleServices {
         return result;
     }
 
-    public String saveLanguageRole(String name) {
-        if (name == null || name.trim().isEmpty()) {
-            return "El role del lenguaje no puede estar vacío.";
-        }
-        if (languageRoleRepo.findLanguageRoleByLanguageRole(name).size() >= 1) {
-            return "Ya existe un rol con ese nombre.";
-        }
-       LanguageRole languageRole = new LanguageRole();
-        languageRole.setLanguageRole(name);
-        languageRoleRepo.save(languageRole);
-        return null;
-    }
-
-
     public LanguageRole findLanguageRoleById(Integer id) {
         return languageRoleRepo.findById(id).get();
     }
@@ -56,26 +45,41 @@ public class LanguageRoleServices {
         return languageRoleRepo.findLanguageRoleByLanguageRole(languageRoleSearch);
     }
 
-    public String deleteLanguageRole(Integer id) {
+    public String saveLanguageRole(String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Crear roles del idioma");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        if (name == null || name.trim().isEmpty()) return "El role del lenguaje no puede estar vacío.";
+        if (languageRoleRepo.findLanguageRoleByLanguageRole(name).size() >= 1)
+            return "Ya existe un rol con ese nombre.";
+        LanguageRole languageRole = new LanguageRole();
+        languageRole.setLanguageRole(name);
+        languageRoleRepo.save(languageRole);
+        return null;
+    }
+
+
+    public String deleteLanguageRole(Integer id, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Eliminar roles del idioma");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
         try {
             languageRoleRepo.deleteById(id);
-            return "Ok";
-        }catch (Exception e) {
-            return "Error";
-        }
-
-    }
-
-    public LanguageRole updateLanguageRole(Integer id, String name) {
-        Optional<LanguageRole> existingLanguageRole = languageRoleRepo.findById(id);
-        if (existingLanguageRole.isPresent()) {
-            LanguageRole updatedLanguageRole = existingLanguageRole.get();
-            updatedLanguageRole.setLanguageRole(name);
-            languageRoleRepo.save(updatedLanguageRole);
-            return updatedLanguageRole;
-        } else {
             return null;
+        } catch (Exception e) {
+            return "Ha habido un error al eliminar el rol del lenguaje";
         }
+
     }
 
+    public String updateLanguageRole(Integer id, String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Modificar roles del idioma");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        List<LanguageRole> sameName = languageRoleRepo.findLanguageRoleByLanguageRole(name);
+        if (!sameName.isEmpty()) return "Ya existe un registro con ese nombre";
+        Optional<LanguageRole> existingLanguageRole = languageRoleRepo.findById(id);
+        if (existingLanguageRole.isEmpty()) return "No existe esa rol";
+        LanguageRole updatedLanguageRole = existingLanguageRole.get();
+        updatedLanguageRole.setLanguageRole(name);
+        languageRoleRepo.save(updatedLanguageRole);
+        return null;
+    }
 }
