+++ src/main/java/com/esliceu/movies/services/LanguageServices.java
@@ -1,5 +1,7 @@
 package com.esliceu.movies.services;
+
 import com.esliceu.movies.models.Language;
+import com.esliceu.movies.models.Person;
 import com.esliceu.movies.repos.LanguageRepo;
 import com.google.gson.Gson;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +17,8 @@ import java.util.stream.Collectors;
 public class LanguageServices {
     @Autowired
     LanguageRepo languageRepo;
+    @Autowired
+    PermissionsServices permissionsServices;
 
     public Page<Language> findAllLanguages(Pageable pageable) {
         return languageRepo.findAll(pageable);
@@ -31,24 +35,6 @@ public class LanguageServices {
         return result;
     }
 
-    public String saveLanguage(String name, String code) {
-        if (name == null || name.trim().isEmpty()) {
-            return "El lenguaje no puede estar vacío.";
-        }
-        if (code == null || code.trim().isEmpty()) {
-            return "El código no puede estar vacío.";
-        }
-        if (languageRepo.findLanguageByLanguageName(name).size() >= 1) {
-            return "Ya existe un lenguaje con ese nombre.";
-        }
-        Language language = new Language();
-        language.setLanguageName(name);
-        language.setLanguageCode(code);
-        languageRepo.save(language);
-        return null;
-    }
-
-
     public Language findLanguageById(Integer id) {
         return languageRepo.findById(id).get();
     }
@@ -57,31 +43,43 @@ public class LanguageServices {
         return languageRepo.findLanguageByLanguageName(languageSearch);
     }
 
-    public String deleteLanguage(Integer id) {
-        try {
-            languageRepo.deleteById(id);
-            return "Ok";
-        }catch (Exception e) {
-            return "Error";
-        }
-
+    public String saveLanguage(String name, String code, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Crear idiomas");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        if (name == null || name.trim().isEmpty()) return "El lenguaje no puede estar vacío.";
+        if (code == null || code.trim().isEmpty()) return "El código no puede estar vacío.";
+        if (languageRepo.findLanguageByLanguageName(name).size() >= 1) return "Ya existe un lenguaje con ese nombre.";
+        Language language = new Language();
+        language.setLanguageName(name);
+        language.setLanguageCode(code);
+        languageRepo.save(language);
+        return null;
     }
 
-    public Language updateLanguage(Integer id, String name, String code) {
-        Optional<Language> existingLanguage = languageRepo.findById(id);
-        if (existingLanguage.isPresent()) {
-            Language updatedLanguage = existingLanguage.get();
-            if (name != null && !name.isEmpty()) {
-                updatedLanguage.setLanguageName(name);
-            }
-            if (code != null && !code.isEmpty()) {
-                updatedLanguage.setLanguageCode(code);
-            }
-            languageRepo.save(updatedLanguage);
-            return updatedLanguage;
-        } else {
+
+    public String deleteLanguage(Integer id, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Eliminar idiomas");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        try {
+            languageRepo.deleteById(id);
             return null;
+        } catch (Exception e) {
+            return "Ha habido un error al eliminar el idioma";
         }
+    }
+
+    public String updateLanguage(Integer id, String name, String code, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Modificar idiomas");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        List<Language> sameName = languageRepo.findLanguageByLanguageName(name);
+        if (!sameName.isEmpty()) return "Ya existe un registro con ese nombre";
+        Optional<Language> existingLanguage = languageRepo.findById(id);
+        if (existingLanguage.isEmpty()) return  "No existe esa idioma";
+        Language updatedLanguage = existingLanguage.get();
+        if (name != null && !name.isEmpty()) updatedLanguage.setLanguageName(name);
+        if (code != null && !code.isEmpty()) updatedLanguage.setLanguageCode(code);
+        languageRepo.save(updatedLanguage);
+        return null;
     }
 
 }
